streamlit_app/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
youtube-transcript-api
deep-translator
python-dotenv
openai
streamlit